Extract shared Vue component accessor.
